padroniza arquivos que realizam alterações nos comentários
